requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq
 langchain-text-splitters
 faiss-cpu
 langchain-chroma
 langchain_huggingface
 arxiv
 wikipedia
langgraph
langchainhub
SQLAlchemy
mysql-connector-python
validators == 0.28.1
youtube_transcript_api
unstructured
huggingface_hub
gradio
pinecone-client
langchain-pinecone
pydantic
langchain-nvidia-ai-endpoints
sentence-transformers